+++ riprip_core/src/disc.rs
@@ -82,7 +82,7 @@ impl fmt::Display for Disc {
 			writeln!(
 				f,
 				"\x1b[2m{total:02}  {:>6}                    DATA TRACK\x1b[0m",
-				self.toc.data_sector().unwrap_or_default().saturating_sub(CD_LEADIN)
+				self.toc.data_sector().unwrap_or_default().saturating_sub(u32::from(CD_LEADIN))
 			)?;
 		}
 
@@ -107,7 +107,7 @@ impl fmt::Display for Disc {
 			writeln!(
 				f,
 				"\x1b[2m{total:02}  {:>6}                    DATA TRACK\x1b[0m",
-				self.toc.data_sector().unwrap_or_default().saturating_sub(CD_LEADIN)
+				self.toc.data_sector().unwrap_or_default().saturating_sub(u32::from(CD_LEADIN))
 			)?;
 		}
 
@@ -115,7 +115,7 @@ impl fmt::Display for Disc {
 		writeln!(
 			f,
 			"\x1b[2m{CD_LEADOUT_LABEL}  {:>6}                      LEAD-OUT",
-			self.toc.leadout().saturating_sub(CD_LEADIN),
+			self.toc.leadout().saturating_sub(u32::from(CD_LEADIN)),
 		)?;
 
 		// Close it off!
